Split GraphBasedNodeHandler into separate class.
